+++ CHANGELOG.md
@@ -374,3 +374,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix small orb icons followed by a period in tooltips (kiooeht)
 * Refactor CustomBottleRelic to be an interface (kiooeht)
 * Render custom card tooltips in SingleCardViewPopup (kiooeht)
+* Fix Red/Green/Blue/Colorless Cards modifiers to work with modded characters (kiooeht)